move minimizer to new folder
