+++ plugins/optimizer/minimizer/scipy_minimizer/schemas.py
